feat: added quick edits for media details pages
