+++ services/media/workflows/package.json
@@ -27,8 +27,8 @@
   ],
   "dependencies": {
     "@apollo/client": "^3.7.10",
-    "@axinom/mosaic-managed-workflow-integration": "0.18.0",
-    "@axinom/mosaic-ui": "0.50.0",
+    "@axinom/mosaic-managed-workflow-integration": "0.19.0",
+    "@axinom/mosaic-ui": "0.51.0",
     "apollo-upload-client": "^14.1.3",
     "clsx": "^1.2.1",
     "env-cmd": "^10.1.0",
@@ -39,7 +39,7 @@
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@axinom/mosaic-portal": "0.34.0",
+    "@axinom/mosaic-portal": "0.35.0",
     "@dbeining/react-atom": "^4.1.21",
     "@graphql-codegen/cli": "^2.13.4",
     "@graphql-codegen/typescript": "^2.7.3",
