+++ services/media/workflows/src/Stations/Movies/MovieExplorerBase/MovieExplorer.tsx
@@ -29,6 +29,7 @@ import {
   useUnpublishMovieMutation,
 } from '../../../generated/graphql';
 import { PublishStatusStateMap } from '../../../Util/PublishStatusStateMap/PublishStatusStateMap';
+import { MovieDetailsQuickEdit } from '../MovieDetails/MovieDetailsQuickEdit';
 import { useMoviesFilters } from './MovieExplorer.filters';
 import { MovieData, MovieExplorerProps } from './MovieExplorer.types';
 
@@ -203,6 +204,9 @@ export const MovieExplorer: React.FC<MovieExplorerProps> = (props) => {
           filterOptions={filterOptions}
           defaultSortOrder={{ column: 'updatedDate', direction: 'desc' }}
           inlineMenuActions={generateInlineMenuActions}
+          quickEditRegistrations={[
+            { component: <MovieDetailsQuickEdit />, label: 'Movie Details' },
+          ]}
         />
       );
     case 'SelectionExplorer':
